ns: Move tests out of the mac package
