+++ pkg/networkserver/mac/mac_internal_test.go
@@ -14,395 +14,4 @@
 
 package mac
 
-import (
-	"testing"
-
-	"github.com/smartystreets/assertions"
-	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
-	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
-)
-
-func TestHandleMACResponseBlock(t *testing.T) {
-	t.Parallel()
-
-	for _, tc := range []struct {
-		Name string
-
-		CID          ttnpb.MACCommandIdentifier
-		AllowMissing bool
-		F            func(*assertions.Assertion, *ttnpb.MACCommand) error
-		Commands     []*ttnpb.MACCommand
-
-		Result      []*ttnpb.MACCommand
-		AssertError func(*assertions.Assertion, error)
-	}{
-		{
-			Name: "LinkADRReq",
-
-			CID:          ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-			AllowMissing: false,
-			F: func(a *assertions.Assertion, cmd *ttnpb.MACCommand) error {
-				a.So(cmd, should.Resemble, &ttnpb.MACCommand{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				})
-				return nil
-			},
-			Commands: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-			},
-
-			Result: []*ttnpb.MACCommand{},
-		},
-		{
-			Name: "LinkADRReq+LinkADRReq",
-
-			CID:          ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-			AllowMissing: false,
-			F: func(a *assertions.Assertion, cmd *ttnpb.MACCommand) error {
-				a.So(cmd, should.Resemble, &ttnpb.MACCommand{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				})
-				return nil
-			},
-			Commands: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-			},
-
-			Result: []*ttnpb.MACCommand{},
-		},
-		{
-			Name: "LinkADRReq+DlChannelReq",
-
-			CID:          ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-			AllowMissing: false,
-			F: func(a *assertions.Assertion, cmd *ttnpb.MACCommand) error {
-				a.So(cmd, should.Resemble, &ttnpb.MACCommand{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				})
-				return nil
-			},
-			Commands: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-
-			Result: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-		},
-		{
-			Name: "DlChannelReq+LinkADRReq",
-
-			CID:          ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-			AllowMissing: false,
-			F: func(a *assertions.Assertion, cmd *ttnpb.MACCommand) error {
-				a.So(cmd, should.Resemble, &ttnpb.MACCommand{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				})
-				return nil
-			},
-			Commands: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-			},
-
-			Result: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-		},
-		{
-			Name: "LinkADRReq+LinkADRReq+DlChannelReq",
-
-			CID:          ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-			AllowMissing: false,
-			F: func(a *assertions.Assertion, cmd *ttnpb.MACCommand) error {
-				a.So(cmd, should.Resemble, &ttnpb.MACCommand{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				})
-				return nil
-			},
-			Commands: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-
-			Result: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-		},
-		{
-			Name: "DlChannelReq+LinkADRReq+LinkADRReq+DlChannelReq",
-
-			CID:          ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-			AllowMissing: false,
-			F: func(a *assertions.Assertion, cmd *ttnpb.MACCommand) error {
-				a.So(cmd, should.Resemble, &ttnpb.MACCommand{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				})
-				return nil
-			},
-			Commands: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_LINK_ADR,
-					Payload: &ttnpb.MACCommand_LinkAdrReq{
-						LinkAdrReq: &ttnpb.MACCommand_LinkADRReq{
-							DataRateIndex:      1,
-							TxPowerIndex:       2,
-							ChannelMask:        []bool{false, true, false},
-							ChannelMaskControl: 3,
-							NbTrans:            4,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-
-			Result: []*ttnpb.MACCommand{
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-				{
-					Cid: ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
-					Payload: &ttnpb.MACCommand_DlChannelReq{
-						DlChannelReq: &ttnpb.MACCommand_DLChannelReq{
-							ChannelIndex: 1,
-							Frequency:    2,
-						},
-					},
-				},
-			},
-		},
-	} {
-		tc := tc
-		t.Run(tc.Name, func(t *testing.T) {
-			t.Parallel()
-
-			a := assertions.New(t)
-			rest, err := handleMACResponseBlock(tc.CID, tc.AllowMissing, func(cmd *ttnpb.MACCommand) error {
-				return tc.F(a, cmd)
-			}, tc.Commands...)
-			if tc.AssertError == nil {
-				a.So(err, should.BeNil)
-				a.So(rest, should.Resemble, tc.Result)
-			} else {
-				tc.AssertError(a, err)
-			}
-		})
-	}
-}
+var HandleMACResponseBlock = handleMACResponseBlock
